% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgmguru-functions-docs.R
\name{find_local_maxima}
\alias{find_local_maxima}
\title{Find Local Maxima in Glucose Time Series}
\arguments{
\item{df}{A dataframe containing CGM data with columns: id, time, glucose}
}
\value{
A dataframe containing identified local maxima with their properties
}
\description{
Identifies local maxima (peaks) in glucose concentration time series data.
Useful for identifying potential glycemic event peaks.
}
\examples{
\dontrun{
# Find local maxima
maxima <- find_local_maxima(cgm_data)
}
}
